fix name stuff and standarize the code
