+++ scripts/sample_extraction.py
@@ -1,14 +1,13 @@
 """
 Script to manualy cut the target sample of an image.
-You can do it calling extract_sample() and selecting the desire image.
+You can do it calling extract_sample(img) passing the 
+image you want to cut.
 """
 import cv2
 import numpy as np
-import api_fm
 import math
 
-# Variables globales
-
+TOKEN = "[EXTRACT]"
 
 def draw_circles_and_lines(frame, r, c1, c2, c3, c4):
     cv2.circle(frame, c1, r, RED, -1) 
@@ -20,7 +19,8 @@ def draw_circles_and_lines(frame, r, c1, c2, c3, c4):
     cv2.line(frame, c3, c4, BLUE)
     cv2.line(frame, c3, c1, BLUE) 
 
-#fórmula pitagórica para calcular diferencia entre el punto del click y el de algún otro punto
+# Pythagorean formula for calculating the difference between 
+# the click point and some other point.
 def dif_circle(center, x, y):
     dif_x = abs(x - center[0])
     dif_y = abs(y - center[1])
@@ -61,7 +61,7 @@ def mouse(event,x,y, flags, params):
         move_circle_3 = False
         move_circle_4 = False
 
-def extract_sample():
+def extract_sample(img):
     global RED, BLUE, GREEN, move_circle_1, move_circle_2, move_circle_3, move_circle_4
     global radius, bg, bg_original, bg_size, r1_center, r2_center, r3_center, r4_center
     move_circle_1 = False
@@ -72,47 +72,43 @@ def extract_sample():
     RED = [0, 0, 255]
     GREEN = [0, 255, 0]
 
-    bg = api_fm.load_image()
+    img = cv2.resize(img, (int(img.shape[1]*0.2),int(img.shape[0]*0.2)))
+    bg = img
 
     bg_original = bg.copy()
 
     bg_size = bg.shape
 
-    #radio de los puntos visibles en la imagen
+    # radius of the visible points in the image
     radius = 7
 
-    #Estos son los 4 puntos (en orden son A, B, C y D), son variables globales
-    #En teoría siempre son A, B, C y D en ese orden, a menos que alguien deforme el cuadrilátero visible
-    #(o lo invierta), estos casos son mejorables pero lo haré a futuro (para impedir que un punto cruze
-    # una línea)
-    #Tienen valores iniciales equivalentes a las 4 esquinas de la imagen y se van modificando con el código,
-    #si están muy esquinadas inicialmente se pueden correr algunos píxeles hacia adentro en la pos. inicial
+    # The points r1,r2,r4,r3 _center are the corners of the rectangle, the are
+    # global vars and it must be in these order (1,2,4,3).
+    # TODO prevent the inversion of the point order.
+    # Initially the 4 points are in the center of the image.
 
-    r1_center = (0,0)
-    r2_center = (0,bg_size[0])
-    r4_center = (bg_size[1],bg_size[0])
-    r3_center = (bg_size[1],0)
+    r1_center = (bg_size[1]//4, bg_size[0]//4)
+    r2_center = (bg_size[1]//4, bg_size[0]*3//4)
+    r4_center = (bg_size[1]*3//4, bg_size[0]*3//4)
+    r3_center = (bg_size[1]*3//4, bg_size[0]//4)
 
     draw_circles_and_lines(bg, radius, r1_center, r2_center, r3_center, r4_center)    
 
     cv2.namedWindow('draw')
     cv2.setMouseCallback('draw', mouse)
-
-    #Por ahora, para matar al código hay que apretar Esc (si se aprieta la X se vuelve a poner la imagen,
-    # y por razones desconocidas no se mueven los círculos, pero en teoría no se debería hacer eso XD, 
-    # error a arreglar a futuro)
-
+    print(f"{TOKEN} Use 's' to save or 'r' to reset the cut, 'Esc' for exit.")
+    
     while True:
-
+        
         cv2.imshow('draw', bg)
         k = cv2.waitKey(1)
 
-        #waiting for esc to exit
+        # if 'Esc' is pressed, the cuting stops.
         if k == 27 & 0xFF:
             cv2.destroyAllWindows()
             return
 
-        #Se ejecuta el corte al presionar "s"
+        # if 's' is pressed, the img was cut in the rectangle area.
         elif k == ord("s"):
             refPt = [r1_center, r2_center, r4_center, r3_center]
 
@@ -130,7 +126,7 @@ def extract_sample():
                                     [0, maxHeight - 1],
                                     [maxWidth - 1, maxHeight - 1],
                                     [maxWidth - 1, 0]])
-            # Se contruye la persepectiva y se corta sobre un clon de la imagen original.
+            # The perspective is built and cut on a clone of the original image.
             M = cv2.getPerspectiveTransform(input_pts,output_pts)
             out = cv2.warpPerspective(bg_original,M,(maxWidth, maxHeight),flags=cv2.INTER_LINEAR)
             cv2.imshow("Image cut", out)
@@ -138,4 +134,9 @@ def extract_sample():
             cv2.destroyAllWindows()
             return out
 
-            #si
+        # if 'r' is pressed, the rectangle return to the original position.
+        elif k == ord('r'):
+            r1_center = (0,0)
+            r2_center = (0,bg_size[0])
+            r4_center = (bg_size[1],bg_size[0])
+            r3_center = (bg_size[1],0)
